aoc2023: optimize day 7
Replaces the overengineered search algorithm with a simpler
implementation that simply adds the jokers to the most common card in
the hand.

Use zero-based indexing for Hand.

Pre-count the number of cards in each hand.

Replace Card'Value with a function that maps Characters directly.

Overall this results in a little over a 2x total speedup for a 2.5M
input, from 600ms to 270ms. The majority of remaining time is spent in
the Ada standard library loading the file line-by-line and sorting the
vectors (~220ms).