aoc2023-util: simplify Direction subtypes
Reorders the values of Direction so that the subtypes Cardinal_Direction
and Ordinal_Direction can be specified as a subrange of Direction
instead of using Static_Predicate. Static_Predicate prevents the use of
the subtype as an array range because the values aren't contiguous.

With the new representation, one can write `array (Cardinal_Direction)`
to get an array with only 4 valid indices.