aoc2024-06: optimize
Reduces runtime by roughly a factor of 10x in both Debug and Release
modes from 3 seconds to 300 milliseconds.

Part 1 and part 2 solvers are split since they need differing amounts of
information.

Part 1 now records a list of distinct, visited locations and returns it
for use by the part 2 solver.

Only part 1 needs to count the number of visited locations. Instead of
counting in linear time, it now counts as new locations are encountered.

The 3-way cache in part 2 now uses only Cardinal_Directions instead of
the full range of Directions. This probably has a minor effect on cache
usage.

Obstacle placement in part 2 now only uses locations the guard visited
in part 1. This significantly reduces the number of simulations.