aoc2024-09: optimize
Debug build:   2530ms -> 14ms
Release build: 920ms  -> 11ms

The main optimization here is to use an array of min-heaps to quickly
find the next span of free blocks. This allows us to find a free block
to move a file to in constant time.

Extracting from the min-heap is O(log(N)) - but N is the number of free
block spans with the same width, which is considerably less than the
number of blocks in the input.