aoc2024-16: second optimization pass
Reduces runtime from 930ms to 35ms. The main optimization this time is
reducing vector copies.

Instead of storing the full path as a vector in the path state, we store
the path externally and refer to them by an integer index instead. This
design also makes it easy to deduplicate paths. Instead of copying the
whole path from start to finish like before, we only need to allocate a
new position and parent it to the path we came from.

Additionally, we no longer copy tile position vectors when considering
edges in the graph. We introduce `Junction_Out_Edge_Ptr` for this
purpose, as accessing the edge through an access type doesn't copy its
vector.